Fixed bugs in !promo, optimized video download, and playback.
Fixed a bug where a video from !promo would enter an infinite loop if a number is not used at the end of the name file. Also, resolved a bug where !promo would not display the video the first 1/2 times it is used. Now, the page waits for the user to fully download the clip before playing it, ensuring that people with a poor connection don't experience a cut-off video.